pyproject.toml
[project]
name = "victron-mqtt-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "paho-mqtt>=2.1.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-i18n[yaml]>=0.3.9",
    "tapo>=0.8.4",
    "tzlocal>=5.3.1",
    "uvicorn>=0.35.0",
]
